Add meson and use catch2
[why]
Because we use meson to build stuff and we need this to incorporate it
into our projects.
Catch2 is used by GUL etc and so the tests should base on that to avoid
a new dependency.

[how]
Catch2's header file must be copied here manually before building.

[note]
Also add a very simple usage example.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>